Release/0.7.1 (#229)
Improve Value Converters and issues

- Created reference to a FieldExpression in LiteralExpression, AssignmentExpression, and InsertExpression to have a direct correlation between the source FieldExpression and the using Expression.
- Deprecated FieldExpression from AssemblyContext.  As FieldExpression is now part of the creation of Expressions related to the FieldExpression, using the Push/Pop methods for a FieldExpression during assembly were no longer needed.
- Refactored parameter builders to provide consistency across all creation methods by always returning a ParameterizedExpression.
- Refactored function expressions to "hide" properties via IExpressionProvider implementation.  This ensures no confusion in the API when used (i.e. db.fx.DatePart(year, date1, date1).DatePart is confusing)
- Cosmetic change for assembling SQL statements: removed extraneous indentation while building queries containing offset and limit specifications.
- [issue #225](https://github.com/HatTrickLabs/dbExpression/issues/225): Scaffolding generation now honors ignore configuration option.  Previous to this release, all views for a database would be scaffolded.

+++ test/HatTrick.DbEx.MsSql.Test/_TypeCode/GenderType.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 namespace DbEx.Data
 {
     #region GenderType
-    public enum GenderType : int
+    public enum GenderType : long
 	{
 		[Display(Name = "Mail", Description = "Male")]
 		Male = 1,
